@@ -1,8 +1,7 @@
 package com.sivalabs.jblogger.domain;
 
-import lombok.Data;
-
 import java.util.List;
+import lombok.Data;
 
 @Data
 public class PostsResponse {
