@@ -3,20 +3,21 @@ package com.sivalabs.jblogger.domain;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public enum TimePeriod
-{
-	TODAY, YESTERDAY, WEEK, MONTH, ALL_TIME;
-	
-	private static final Logger logger = LoggerFactory.getLogger(TimePeriod.class);
-	
-	public static TimePeriod fromString(String value) {
-		try
-		{
-			return TimePeriod.valueOf(value);
-		} catch (Exception e)
-		{
-			logger.error(e.getMessage(),e);
-			return TimePeriod.TODAY;
-		}
-	}
+public enum TimePeriod {
+    TODAY,
+    YESTERDAY,
+    WEEK,
+    MONTH,
+    ALL_TIME;
+
+    private static final Logger logger = LoggerFactory.getLogger(TimePeriod.class);
+
+    public static TimePeriod fromString(String value) {
+        try {
+            return TimePeriod.valueOf(value);
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            return TimePeriod.TODAY;
+        }
+    }
 }
