@@ -1,32 +1,26 @@
 package com.sivalabs.jblogger.entities;
 
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.time.LocalDateTime;
-
-import javax.persistence.*;
 import lombok.Data;
 
 @Entity
-@Table(name="pageviews")
+@Table(name = "pageviews")
 @Data
-public class PageView implements Serializable
-{
-	private static final long serialVersionUID = 1L;
-	
-	@Id
-	@SequenceGenerator(name="pageview_id_generator", sequenceName="pageview_id_seq", initialValue = 100, allocationSize=1)
-	@GeneratedValue(generator = "pageview_id_generator")
-	private Long id;
-	
-	private String url;
-	
-	private String referrer;
-	
-	@Column(name="visit_time")
-	private LocalDateTime visitTime = LocalDateTime.now();
-	
-	@ManyToOne
-	@JoinColumn(name="post_id")
-	private Post post;
+public class PageView implements Serializable {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
+    private String url;
+
+    private String referrer;
+
+    @Column(name = "visit_time")
+    private LocalDateTime visitTime = LocalDateTime.now();
+
+    @ManyToOne
+    @JoinColumn(name = "post_id")
+    private Post post;
 }
