@@ -1,12 +1,8 @@
 package com.sivalabs.jblogger.repositories;
 
+import com.sivalabs.jblogger.entities.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.sivalabs.jblogger.entities.Comment;
-
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long>
-{
-
-}
+public interface CommentRepository extends JpaRepository<Comment, Long> {}
