@@ -1,6 +1,8 @@
 package com.sivalabs.jblogger.security;
 
+import com.sivalabs.jblogger.entities.User;
 import com.sivalabs.jblogger.repositories.UserRepository;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -8,29 +10,23 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sivalabs.jblogger.entities.User;
-
-import java.util.Optional;
-
 @Service("userDetailsService")
 @Transactional
-public class SecurityUserDetailsService implements UserDetailsService
-{
-	private UserRepository userRepository;
+public class SecurityUserDetailsService implements UserDetailsService {
+    private UserRepository userRepository;
 
-	@Autowired
-	public SecurityUserDetailsService(UserRepository userRepository) {
-		this.userRepository = userRepository;
-	}
+    @Autowired
+    public SecurityUserDetailsService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
-	@Override
-	public UserDetails loadUserByUsername(String userName) {
-		Optional<User> user = userRepository.findByEmail(userName);
-		if(user.isPresent()){
-			return new AuthenticatedUser(user.get());
-		} else {
-			throw new UsernameNotFoundException("Email "+userName+" not found");
-		}
-	}
-
+    @Override
+    public UserDetails loadUserByUsername(String userName) {
+        Optional<User> user = userRepository.findByEmail(userName);
+        if (user.isPresent()) {
+            return new AuthenticatedUser(user.get());
+        } else {
+            throw new UsernameNotFoundException("Email " + userName + " not found");
+        }
+    }
 }
